refactor: Re-work LocalStack extensions into modules
