+++ lambda/bootstrap/simple.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package main
+package bootstrap
 
 import (
 	"fmt"
