+++ lambda/logging/egress_api.go
@@ -1,4 +1,4 @@
-package main
+package logging
 
 import (
 	"io"
