+++ lambda/tracing/tracer.go
@@ -1,8 +1,9 @@
-package main
+package tracing
 
 import (
 	"context"
 	"encoding/json"
+
 	"go.amzn.com/lambda/appctx"
 	"go.amzn.com/lambda/interop"
 )
